reorganize dialog components
