+++ frontend/src/components/dialogs/base/BaseDialog.module.scss
