+++ frontend/src/components/dialogs/consent/ConsentDialog.module.scss
