+++ frontend/src/components/dialogs/snapshot/SnapshotDialog.module.scss
