+++ frontend/src/components/requirements/ProjectRequirements.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ActionDialog } from "../action-dialog/ActionDialog";
+import { ActionDialog } from "../dialogs/action/ActionDialog";
 import { useGetProjectRequirements } from "../../hooks/use-api";
 import { ReactElement } from "react";
 import { ProjectRequirement, ProjectRequirementType } from "@flowser/shared";
